TEIID-2707: proactive buffering on an enhanced sort merge left outer join results in an assertionerror
git-svn-id: http://svn.jboss.org/repos/teiid/branches/7.7.x@4612 4ebdeab9-025a-0410-8b34-84821ecf4c0f

+++ engine/src/main/java/org/teiid/query/processor/relational/SortUtility.java
@@ -31,11 +31,11 @@ import java.util.TreeSet;
 
 import org.teiid.common.buffer.BlockedException;
 import org.teiid.common.buffer.BufferManager;
+import org.teiid.common.buffer.BufferManager.BufferReserveMode;
+import org.teiid.common.buffer.BufferManager.TupleSourceType;
 import org.teiid.common.buffer.IndexedTupleSource;
 import org.teiid.common.buffer.TupleBuffer;
 import org.teiid.common.buffer.TupleSource;
-import org.teiid.common.buffer.BufferManager.BufferReserveMode;
-import org.teiid.common.buffer.BufferManager.TupleSourceType;
 import org.teiid.core.TeiidComponentException;
 import org.teiid.core.TeiidProcessingException;
 import org.teiid.core.util.Assertion;
@@ -301,7 +301,7 @@ public class SortUtility {
             
             TupleBuffer merged = createTupleBuffer();
 
-            int desiredSpace = activeTupleBuffers.size() * schemaSize;
+            int desiredSpace = (int)(activeTupleBuffers.size() * (long)schemaSize);
             int reserved = Math.min(desiredSpace, Math.max(2*schemaSize, this.bufferManager.getMaxProcessingSize()));
             bufferManager.reserveBuffers(reserved, BufferReserveMode.FORCE);
             if (desiredSpace > reserved) {
